Extracted compiling example groups for each QuickSpec subclasses from QuickTestObservation into QuickTestExamplesLoader
